row_performance Unit tests pass

+++ bi_etl/components/row/row_iteration_header.py
@@ -3,7 +3,6 @@ Created on May 26, 2015
 
 @author: woodd
 """
-import functools
 from operator import attrgetter
 from typing import Union, List, Iterable
 
@@ -43,19 +42,25 @@ class RowIterationHeader(object):
         self.primary_key = primary_key
         self.parent = parent
         self._actions_to_next_headers = dict()
-        self.columns_in_order = list()
+        self._columns_in_order = list()
         self._columns_positions = dict()
         self.columns_frozen = False
         self._cached_column_set = None
         self._cached_positioned_column_set = None
+        self._action_position = None
+        self.action_id = None
+        self.action_count = 0
 
-    def get_next_header(self, action: str) -> 'RowIterationHeader':
+    def get_next_header(self, action: Union[str, tuple], start_empty: bool = False) -> 'RowIterationHeader':
         """
         Get the next header after performing a manipulation on the set of columns.
 
         Parameters
         ----------
-        action
+        action:
+            A hashable action ID
+        start_empty:
+            Should the new header start empty (vs transferring the columns)
 
         Returns
         -------
@@ -63,15 +68,36 @@ class RowIterationHeader(object):
         """
         if action not in self._actions_to_next_headers:
             new_header = RowIterationHeader(all_rows_same_columns=self.all_rows_same_columns,
-                                            logical_name= (self.logical_name + '.' + action),
+                                            logical_name=self.logical_name,
                                             primary_key=self.primary_key,
                                             parent=None
                                             )
-            new_header.columns_in_order = self.columns_in_order
-            new_header._columns_positions = self._columns_positions
+            new_header.action_count = self.action_count + 1
+            new_header.action_id = action
+            if not start_empty:
+                new_header._columns_in_order = self._columns_in_order.copy()
+                new_header._columns_positions = self._columns_positions.copy()
 
             self._actions_to_next_headers[action] = new_header
         return self._actions_to_next_headers[action]
+
+    def __repr__(self):
+        return '{cls}(id={id},logical_name={logical_name},primary_key={pk}'.format(
+            cls=self.__class__.__name__,
+            id=self.iteration_id,
+            logical_name=self.logical_name,
+            pk=self.primary_key,
+        )
+
+    def __str__(self):
+        repr(self)
+
+    @property
+    def columns_in_order(self):
+        """
+        A list of the columns of this row in the order they were defined.
+        """
+        return self._columns_in_order.copy()
 
     @property
     def primary_key(self):
@@ -107,7 +133,7 @@ class RowIterationHeader(object):
         See positioned_column_set instead.
         """
         if self._cached_column_set is None:
-            self._cached_column_set = frozenset(self.columns_in_order)
+            self._cached_column_set = frozenset(self._columns_in_order)
         return self._cached_column_set
 
     @property
@@ -121,7 +147,7 @@ class RowIterationHeader(object):
         """
         if self._cached_positioned_column_set is None:
             tpl_lst = list()
-            for key, position in enumerate(self.columns_in_order):
+            for key, position in enumerate(self._columns_in_order):
                 tpl = (key, position)
                 tpl_lst.append(tpl)
             self._cached_positioned_column_set = get_cached_frozen_set(tpl_lst)
@@ -134,7 +160,7 @@ class RowIterationHeader(object):
         self.row_count -= 1
 
     def column_count(self) -> int:
-        return len(self.columns_in_order)
+        return len(self._columns_in_order)
 
     def has_column(self, column_name):
         return column_name in self._columns_positions
@@ -143,7 +169,7 @@ class RowIterationHeader(object):
         return KeyError("{name} has no item {column_name} it does have {cols}"
                         .format(name=self.logical_name,
                                 column_name=column_name,
-                                cols=self.columns_in_order
+                                cols=self._columns_in_order
                                 )
                         )
 
@@ -154,16 +180,31 @@ class RowIterationHeader(object):
             if self.columns_frozen or not allow_create:
                 raise self._key_error(column_name)
             else:
-                new_pos = len(self.columns_in_order)
-                self.columns_in_order.append(column_name)
-                self._columns_positions[column_name] = new_pos
-                self._cached_column_set = None
-                self._cached_positioned_column_set = None
-                position = new_pos
+                position = self.add_column(column_name)
         return position
 
-    @functools.lru_cache()
-    def header_rename_column(self, old_name, new_name, ignore_missing = False) -> int:
+    def row_set_item(self, column_name: str, value, row: 'bi_etl.components.row.row.Row')-> 'RowIterationHeader':
+        if column_name in self._columns_positions:
+            position = self._columns_positions[column_name]
+            row.set_by_zposition(position, value)
+            new_header = self
+        else:
+            # Modification of columns required
+            action = tuple(['+:', column_name])
+            new_header = self.get_next_header(action)
+            new_header.add_row(row)
+            if new_header._action_position is None:
+                new_header._action_position = new_header.add_column(column_name)
+
+            # Protected access is required here since we can't call setitem it calls this method.
+            row._data_values.append(value)
+        return new_header
+
+    def rename_column(self,
+                      old_name: str,
+                      new_name: str,
+                      ignore_missing: bool = False,
+                      no_new_header: bool = False)-> 'RowIterationHeader':
         """
         Rename a column
 
@@ -178,31 +219,46 @@ class RowIterationHeader(object):
         ignore_missing: boolean
             Ignore (don't raise error) if we don't have a column with the name in old_name.
             Defaults to False
+
+        no_new_header:
+            Skip creating a new row header, modify in place.
+            ----------------------
+            BE CAREFUL USING THIS!
+            ----------------------
+            All new rows created with this header will immediately get the new name,
+            in which case you won't want to call this method again.
         """
         assert new_name not in self._columns_positions, "Target column name {} already exists".format(new_name)
         try:
             position = self._columns_positions[old_name]
-            self.columns_in_order[position] = new_name
-            del self._columns_positions[old_name]
-            self._columns_positions[new_name] = position
-            if self.primary_key is not None:
-                try:
-                    pk_position = self.primary_key.index(old_name)
-                    self.primary_key[pk_position] = new_name
-                except ValueError:
-                    pass
-
-            return position
-
+            # Modification of columns required
+            if not no_new_header:
+                action = tuple(['r', old_name, new_name])
+                new_header = self.get_next_header(action)
+                new_header.row_count = self.row_count
+            else:
+                new_header = self
+            if old_name in new_header._columns_positions:
+                new_header._columns_in_order[position] = new_name
+                del new_header._columns_positions[old_name]
+                new_header._columns_positions[new_name] = position
+                if new_header.primary_key is not None:
+                    try:
+                        pk_position = new_header.primary_key.index(old_name)
+                        new_header.primary_key[pk_position] = new_name
+                    except ValueError:
+                        pass
         except (KeyError, ValueError) as e:
             if ignore_missing:
-                return None
+                return self
             else:
                 raise e
+        return new_header
 
     def rename_columns(self,
                        rename_map: Union[dict, List[tuple]],
-                       ignore_missing: bool = False):
+                       ignore_missing: bool = False,
+                       no_new_header: bool = False)-> 'RowIterationHeader':
         """
         Rename many columns at once.
 
@@ -215,51 +271,60 @@ class RowIterationHeader(object):
         ignore_missing
             Ignore (don't raise error) if we don't have a column with the name in old_name.
             Defaults to False
+
+        no_new_header:
+            Skip creating a new row header, modify in place.
+            ----------------------
+            BE CAREFUL USING THIS!
+            ----------------------
+            All new rows created with this header will immediately get the new name,
+            in which case you won't want to call this method again.
         """
-        # Note: subset with rename_map is currently slightly faster
+        new_header = self
         if isinstance(rename_map, dict):
             for k in rename_map.keys():
-                self.rename_column(k, rename_map[k], ignore_missing)
+                new_header = new_header.rename_column(k, rename_map[k],
+                                                      ignore_missing=ignore_missing,
+                                                      no_new_header=no_new_header)
         elif rename_map is not None:  # assume it's a list of tuples
             for (old, new) in rename_map:
-                self.rename_column(old, new, ignore_missing)
+                new_header = new_header.rename_column(old, new,
+                                                      ignore_missing=ignore_missing,
+                                                      no_new_header=no_new_header)
+        return new_header
 
-    def remove_column(self, column_name, ignore_missing = False) -> int:
-        try:
-            position = self._columns_positions[column_name]
-            del self.columns_in_order[position]
-            for following_col in self.columns_in_order[position:]:
-                self._columns_positions[following_col] -= 1
-            return position
-        except KeyError as e:
-            if ignore_missing:
-                return None
-            else:
-                raise e
+    def row_remove_column(self,
+                          column_name: str,
+                          row: 'bi_etl.components.row.row.Row',
+                          ignore_missing: bool=False)-> 'RowIterationHeader':
+        if column_name not in self._columns_positions:
+            if not ignore_missing:
+                raise self._key_error(column_name)
+            return self
+        else:
+            # Modification of columns required
+            action = tuple(['-:', column_name])
+            new_header = self.get_next_header(action)
+            new_header.add_row(row)
+            if new_header._action_position is None:
+                position = new_header._columns_positions[column_name]
+                new_header._action_position = position
+                del new_header._columns_in_order[position]
+                del new_header._columns_positions[column_name]
+                for following_col in new_header._columns_in_order[position:]:
+                    new_header._columns_positions[following_col] -= 1
+                if self.primary_key is not None and column_name in self.primary_key:
+                    self.primary_key.remove(column_name)
+            # Protected access is required here since we can't call __delitem__, it calls this method.
+            del row._data_values[new_header._action_position]
+            return new_header
 
-    def remove_columns(self,
-                       remove_list,
-                       ignore_missing = False):
-        """
-        Remove columns from this row instance.
-
-        Parameters
-        ----------
-        remove_list:
-            A list of column names to remove
-
-        ignore_missing:
-            Ignore (don't raise error) if we don't have a column with a given name
-            Defaults to False
-        """
-        for column_name in remove_list:
-            self.remove_column(column_name, ignore_missing=ignore_missing)
-
-    def subset(self,
-               exclude: Iterable = None,
-               rename_map: Union[dict, List[tuple]] = None,
-               keep_only: Iterable = None,
-              ) -> list:
+    def row_subset(self,
+                   row: 'bi_etl.components.row.row.Row',
+                   exclude: Iterable = None,
+                   rename_map: Union[dict, List[tuple]] = None,
+                   keep_only: Iterable = None,
+                   ) -> 'bi_etl.components.row.row.Row':
         """
         Return a new row instance with a subset of the columns. Original row is not modified
         Excludes are done first, then renames and finally keep_only.
@@ -286,18 +351,56 @@ class RowIterationHeader(object):
             The second item in the list will be the index of that item in the old list.
             etc
         """
-        position_mapping_list = list()
-        old_names = self.columns_in_order.copy()
-        #old_name_positions = self._columns_positions.copy()
-        self.rename_columns(rename_map)
-        for parent_position, old_column_name in enumerate(old_names):
-            remove_column = True
-            new_column_name = self.columns_in_order[parent_position]
-            if old_column_name not in exclude:
-                if keep_only is None or new_column_name in keep_only:
-                    position_mapping_list.append(parent_position)
-                    remove_column = False
-            if remove_column:
-                self.remove_column(new_column_name)
-        return position_mapping_list
+        action_list = ['s:']
 
+        if exclude is not None:
+            action_list.append(tuple(exclude))
+        else:
+            action_list.append(None)
+
+        if keep_only is not None:
+            action_list.append(tuple(keep_only))
+        else:
+            action_list.append(None)
+
+        if isinstance(rename_map, dict):
+            for k in rename_map.keys():
+                action_list.append(tuple([k, rename_map[k]]))
+        elif rename_map is not None:  # assume it's a list of tuples
+            for (old, new) in rename_map:
+                action_list.append(tuple([old, new]))
+
+        action = tuple(action_list)
+        new_iteration_header = self.get_next_header(action)
+        sub_row = row.__class__(iteration_header=new_iteration_header)
+
+        if new_iteration_header._action_position is None:
+            # Do the transformation for the first time
+            self.rename_columns(rename_map, no_new_header=True)
+            old_positions_list = list()
+            sub_pos = 0
+            for parent_position, old_column_name in enumerate(self._columns_in_order):
+                remove_column = True
+                new_column_name = new_iteration_header._columns_in_order[sub_pos]
+                if old_column_name not in exclude:
+                    if keep_only is None or new_column_name in keep_only:
+                        old_positions_list.append(parent_position)
+                        sub_pos += 1
+                        remove_column = False
+                if remove_column:
+                    del new_iteration_header._columns_in_order[sub_pos]
+                    del new_iteration_header._columns_positions[new_column_name]
+            new_iteration_header._action_position = old_positions_list
+        # Build the new row based on the _action_position details
+        for old_pos in new_iteration_header._action_position:
+            # Protected access is required here since we can't call setitem it calls this method.
+            sub_row._data_values.append(row._data_values[old_pos])
+        return sub_row
+
+    def add_column(self, column_name: str) -> int:
+        position = len(self._columns_in_order)
+        self._columns_in_order.append(column_name)
+        self._columns_positions[column_name] = position
+        self._cached_column_set = None
+        self._cached_positioned_column_set = None
+        return position
